requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.11.1
backports-datetime-fromisoformat==2.0.3
billiard==4.2.4
celery==5.6.3
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
Django==5.2.12
django-constance==4.3.5
django-cors-headers==4.9.0
django-modeltranslation==0.20.2
django-multiselectfield==1.0.1
django-ninja==1.6.2
django-unfold==0.81.0
environs==15.0.0
gunicorn==25.3.0
idna==3.11
kombu==5.6.2
marshmallow==4.2.3
packaging==26.0
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
redis==7.4.0
requests==2.33.1
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
